fix(website): fix broken Projects imports in .astro pages and tests
The G-SIZE03 extraction (website-db.ts -> projects-db.ts split) missed
.astro consumers (only .ts/.tsx were scanned) and left a duplicate
listMeetingsForProject/assignMeetingToProject/findProjectByName/
listUnassignedMeetingsForCustomer/getCustomerByEmail block in
website-db.ts that shadowed the moved versions in project-export-db.ts
and project-portal-db.ts.

Changes:
- 13 .astro files (components + pages/admin/*): redirect Projects-domain
  imports (Project, SubProject, ProjectTask, PortalProject, listProjects,
  listAllCustomers, listAdminUsers, listProjectAttachments, etc.) from
  website-db to projects-db / project-export-db, keeping remaining
  website-db imports (TimeEntry, getMeetingsForClient via meetings-db, etc.)
  co-located to minimize line-count deltas on already-oversized files.
- website-db.ts: removed the duplicate Meeting-Projekt-Verknüpfung block
  (listMeetingsForProject, assignMeetingToProject, findProjectByName,
  listUnassignedMeetingsForCustomer, getCustomerByEmail) that duplicated
  project-export-db.ts / project-portal-db.ts; dropped the now-unused
  initMeetingProjectLink import.
- New src/lib/customer-types.ts: extracted the `Customer` interface to a
  neutral leaf module so project-portal-db.ts no longer imports from
  website-db.ts (breaks the S2 import cycle that a naive
  `export { listAllCustomers, listAdminUsers, getCustomerByEmail } from
  './project-portal-db'` re-export in website-db.ts would otherwise
  reintroduce). website-db.ts now re-exports Customer and the three
  portal/customer functions from the neutral module for backward
  compatibility.
- website-db-projects.test.ts / website-db.test.ts: split the projects-db
  import block correctly across website-db (TimeEntry ops, pool,
  claimBrettLinkPost), projects-db (CRUD/attachments/portal), and
  project-export-db (meeting relations) instead of the earlier blind
  sed rewrite that pointed everything at projects-db.

Verified locally:
- npx tsc --noEmit: 0 errors
- npx astro check: 0 errors (0 warnings, pre-existing hints only)
- pnpm lint: 0 errors
- npx vitest run: 327 files / 2638 tests passed (78 pre-existing skips)
- node scripts/code-quality/check.mjs: 0 new/worsened violations (S1/S2 clean)

[T001359]
