+++ website/src/components/portal/KalenderSection.astro
@@ -1,6 +1,6 @@
 ---
 import type { ClientBooking } from '../../lib/caldav';
-import type { PortalProject } from '../../lib/website-db';
+import type { PortalProject } from '../../lib/projects-db';
 
 interface Props {
   bookings: ClientBooking[];
