+++ website/src/components/portal/ProjekteSection.astro
@@ -1,5 +1,5 @@
 ---
-import type { PortalProject } from '../../lib/website-db';
+import type { PortalProject } from '../../lib/projects-db';
 
 interface Props {
   projects: PortalProject[];
