+++ website/src/pages/admin/[clientId].astro
@@ -15,7 +15,7 @@ import ClientContractsPanel from '../../components/admin/ClientContractsPanel.sv
 import ClientQuestionnairesPanel from '../../components/admin/ClientQuestionnairesPanel.svelte';
 import CrmStatusPanel from '../../components/admin/CrmStatusPanel.svelte';
 import ContactHistoryTimeline from '../../components/admin/ContactHistoryTimeline.svelte';
-import { listProjectsForCustomer, getCustomerByEmail } from '../../lib/website-db';
+import { listProjectsForCustomer, getCustomerByEmail } from '../../lib/projects-db';
 import { getSubscriberByEmail } from '../../lib/newsletter-db';
 import { getCustomerProfile, getContactHistory } from '../../lib/customer-crm-db';
