+++ website/src/pages/admin/termine.astro
@@ -4,8 +4,8 @@ import AdminPageHeader from '../../components/admin/ui/AdminPageHeader.svelte';
 import { getSession, getLoginUrl, isAdmin } from '../../lib/auth';
 import { getAllBookings } from '../../lib/caldav';
 import type { AdminBooking } from '../../lib/caldav';
-import { getBookingProjects, listProjects, getBookingInvoices, getBookingLeistungen, getVacationPeriods, getFreeTimeWindows } from '../../lib/website-db';
-import type { Project, BookingInvoiceInfo, VacationPeriod, FreeTimeWindow } from '../../lib/website-db';
+import { getBookingProjects, getBookingInvoices, getBookingLeistungen, getVacationPeriods, getFreeTimeWindows, type BookingInvoiceInfo, type VacationPeriod, type FreeTimeWindow } from '../../lib/website-db';
+import { listProjects, type Project } from '../../lib/projects-db';
 import { getEffectiveLeistungen } from '../../lib/content';
 import type { LeistungCategory } from '../../config/types';
 import AdminBookingModal from '../../components/admin/AdminBookingModal.svelte';
