+++ website/src/pages/api/admin/documents/notify/[id].ts
@@ -1,7 +1,7 @@
 import type { APIRoute } from 'astro';
 import { getSession, isAdmin } from '../../../../../lib/auth';
 import { getDocumentAssignmentById } from '../../../../../lib/documents-db';
-import { getCustomerByEmail } from '../../../../../lib/website-db';
+import { getCustomerByEmail } from '../../../../../lib/projects-db';
 import { logSigningEvent } from '../../../../../lib/signing/audit';
 import nodemailer from 'nodemailer';
